pyproject.toml
[tool.poetry]
name = "rtl_rag_chatbot_api"
version = "1.0.0"
description = "rtl-rag-chatbot-api"
authors = []
readme = 'README.md'

[[tool.poetry.source]]
name = "python-packages"
url = "https://gitlab.com/api/v4/projects/42761560/packages/pypi/simple/"
priority = "supplemental"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
typer = ">=0.7.0"
streamlit = "^1.29.0"
langchain = "^0.0.346"
openai = "^1.3.7"
chromadb = "^0.4.24"
tiktoken = "^0.5.2"
google-cloud-storage = "^2.13.0"
azureopenai = "^0.0.1"
pillow = "^10.1.0"
types-requests = "^2.31.0.20240218"
llama-index = "^0.10.41"
llama-index-embeddings-azure-openai = "^0.1.9"
llama-index-llms-azure-openai = "^0.1.8"
llama-index-vector-stores-chroma = "^0.1.8"
pydantic-settings = "^2.2.1"
fastapi = "^0.109.0"
uvicorn = "^0.27.1"
starlette-exporter = "^0.21.0"

transformers = "^4.30.0"
torch = "^1.13.0"
torchvision = "^0.14.0"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
pandas = "^1.5.3"
opencv-python-headless = "^4.7.0"
pytesseract = "^0.3.10"

cryptography = "^43.0.0"
python-multipart = "^0.0.9"
[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
pytest-cov = "^3.0.0"
pytest-sugar = "^0.9.4"
httpx = "^0.26.0"
isort = "^5.10.1"
black = "^23.3.0"
flake8 = "^6.0.0"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]

[tool.poetry.scripts]
start = "rtl_rag_chatbot_api.app:start"
streamlit = "streamlit run streamlit_app.py"
